requirements.txt
cffi==1.15.1
pycparser==2.21
pycryptodome==3.17
